analysis/requirement.txt
pandas
matplotlib
numpy
spacy
google-cloud-aiplatform
google-genai